Refactor code. Move all job logic to SitempaProcessJob
